Fixed tests and added them to git commit - abstracted rootreducer from App - abstracted custom redux render as test helper - added the running of test suits back to git commit hook
